ai_control_engine/trainers/train_digital_twin.py

import os
import torch
import torch.nn as nn
import torch.optim as optim
from torch.utils.data import DataLoader, TensorDataset
from torch.cuda.amp import GradScaler, autocast
import numpy as np
import yaml
import logging
import time
from typing import Dict, Any, Tuple, Optional
from functools import lru_cache
import warnings
warnings.filterwarnings("ignore")

from models.digital_twin_net import DigitalTwinNet
from utils import save_model, set_seed
from registry.model_registry import log_model_to_registry

# Configure optimized logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s'
)
logger = logging.getLogger("TrainDigitalTwin")
S
# PyTorch optimizations
torch.backends.cudnn.benchmark = True  # Optimize for consistent input sizes
torch.backends.cudnn.deterministic = False  # Allow non-deterministic for speed
if torch.cuda.is_available():
    torch.cuda.empty_cache()


class OptimizedDataGenerator:
    """High-performance dataset generator with vectorized operations."""
    
    def __init__(self, n_samples: int = 10000, n_features: int = 16, 
                 noise_std: float = 0.1, seed: Optional[int] = None):
        self.n_samples = n_samples
        self.n_features = n_features
        self.noise_std = noise_std
        
        if seed is not None:
            np.random.seed(seed)
    
    def generate_dataset(self) -> TensorDataset:
        """Generate synthetic dataset with vectorized operations for speed."""
        # Vectorized generation - much faster than loops
       


class HighPerformanceDataLoader:
    """Optimized DataLoader with automatic performance tuning."""
    
    def __init__(self, dataset: TensorDataset, batch_size: int, shuffle: bool = True):
        # Auto-tune number of workers
 
    
    def __iter__(self):

    
    def __len__(self):



class FastTrainer:
    """Optimized trainer with mixed precision and efficient operations."""
    
    def __init__(self, device: torch.device, use_mixed_precision: bool = True):

    
    def train_epoch(self, model: nn.Module, dataloader: DataLoader, 
                   criterion: nn.Module, optimizer: optim.Optimizer) -> float:
        """Optimized training epoch with mixed precision and efficiency optimizations."""
       
 


@lru_cache(maxsize=2)
def load_config_cached(path: str = "config/ai_engine_config.yaml") -> Dict[str, Any]:
    """Load configuration with caching to avoid repeated file I/O."""



def get_default_config() -> Dict[str, Any]:
    """Get optimized default configuration."""


def create_optimized_model(config: Dict[str, Any], device: torch.device) -> nn.Module:
    """Create and optimize model for training."""



def create_optimized_optimizer(model: nn.Module, config: Dict[str, Any]) -> optim.Optimizer:
    """Create optimized optimizer with better hyperparameters."""



def create_learning_rate_scheduler(optimizer: optim.Optimizer, 
                                 total_steps: int, 
                                 config: Dict[str, Any]) -> Optional[optim.lr_scheduler._LRScheduler]:
    """Create learning rate scheduler for better convergence."""



def efficient_model_saving(model: nn.Module, model_path: str) -> None:
    """Efficient model saving with compression."""



def main():
    """Optimized main training pipeline."""
 


if __name__ == "__main__":
    



