requirements/requirements_quantum.txt
# Quantum frameworks
qiskit>=1.0.0
qutip>=4.7.0

# Machine Learning
scikit-learn>=1.2
tensorflow>=2.12.0
torch>=2.0.0
xgboost>=1.7.5

# Reinforcement Learning (for auto_tuner and pulse_optimizer)
stable-baselines3>=2.1.0
gym>=0.26.2

# Scientific computing
numpy>=1.24
scipy>=1.10
matplotlib>=3.7
pandas>=2.0

# Logging and utility
tqdm>=4.66
joblib>=1.3.0

# Plotting & visualization
seaborn>=0.12
plotly>=5.16.0

# Config and serialization
pyyaml>=6.0
json5>=0.9.14


pytest>=7.3.0
pytest-cov>=4.0
mypy>=1.3
black>=23.3
flake8>=6.0



# ---- Development & Testing ----
pytest
jupyter             # For running and debugging notebooks
notebook            # Backend for Jupyter
ipykernel           # Enable Jupyter kernel interface